+++ src/puppeteer.js
@@ -157,6 +157,8 @@ const getPuppeteerOrThrow = () => {
  *    <li>
  *        The function adds <code>--no-sandbox</code> to <code>args</code> to enable running
  *        headless Chrome in a Docker container on the Apify platform.
+ *        Also it adds <code>--enable-resource-load-scheduler=false</code>
+ *        to make crawling of pages in all tabs run equally fast.
  *    </li>
  * </ul>
  *
@@ -194,6 +196,8 @@ export const launchPuppeteer = (options = {}) => {
 
     optsCopy.args = optsCopy.args || [];
     optsCopy.args.push('--no-sandbox');
+    // TODO: It's not clear that this works, keep eye on https://bugs.chromium.org/p/chromium/issues/detail?id=723233
+    optsCopy.args.push('--enable-resource-load-scheduler=false');
     if (optsCopy.headless == null) {
         // Forcing headless with liveView, otherwise screenshots redirect user to new browser window
         optsCopy.headless = optsCopy.liveView || (process.env[ENV_VARS.HEADLESS] === '1' && process.env[ENV_VARS.XVFB] !== '1');